cygbin/hls_downloader/requirements.txt
m3u8
requests
pyzmq-static
